anomaly_detection/pca/requirements.txt
numpy==1.22.0
Pillow==10.2.0
argparse==1.4.0
